Version 6.7 Community Edition Update

+++ build/scripts/overnight/example_automation_scripts/examples/dcps/Durability/c/standalone/RUN.bat
@@ -1,5 +1,7 @@
 @echo OFF
-rmdir /S /Q C:\tmp\pstore
+IF EXIST C:\tmp\pstore rmdir /S /Q C:\tmp\pstore
+IF EXIST .\tmp\pstore rmdir /S /Q .\tmp\pstore
+
 set EXAMPLE_LANG=sac
 
 echo "===== calling runDurability with sac ====="
